refactor: 정확한 계산을 위한 BigDecimal 적용
https://github.com/next-step/java-lotto/pull/2200#discussion_r763062867

+++ src/main/java/lotto/domain/LottoPrize.java
@@ -1,19 +1,20 @@
 package lotto.domain;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 
 public enum LottoPrize {
-    NOTHING(0, 0),
-    THREE_SAME(5000, 3),
-    FOUR_SAME(50000, 4),
-    FIVE_SAME(1500000, 5),
-    SIX_SAME(2000000000, 6),
+    NOTHING(BigDecimal.valueOf(0), 0),
+    THREE_SAME(BigDecimal.valueOf(5000), 3),
+    FOUR_SAME(BigDecimal.valueOf(50000), 4),
+    FIVE_SAME(BigDecimal.valueOf(1500000), 5),
+    SIX_SAME(BigDecimal.valueOf(2000000000), 6),
     ;
 
-    private final int prizeMoney;
+    private final BigDecimal prizeMoney;
     private final long sameCount;
 
-    LottoPrize(int prizeMoney, int sameCount) {
+    LottoPrize(BigDecimal prizeMoney, int sameCount) {
         this.prizeMoney = prizeMoney;
         this.sameCount = sameCount;
     }
@@ -30,7 +31,7 @@ public enum LottoPrize {
         return this.sameCount == sameCount;
     }
 
-    public int prizeMoney() {
+    public BigDecimal prizeMoney() {
         return prizeMoney;
     }
